Add HttpStatus Codes using ResponseEntity

+++ src/main/java/com/example/bookStore/service/BookStoreService.java
@@ -4,8 +4,11 @@ import com.example.bookStore.entity.BookEntry;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import com.example.bookStore.Repository.BookStoreRepository;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,23 +18,38 @@ public class BookStoreService {
     @Autowired
     private BookStoreRepository bookStoreRepository;
 
-    public void saveEntry(BookEntry bookEntry){
-        bookStoreRepository.save(bookEntry);
+    public ResponseEntity<BookEntry> saveEntry(BookEntry bookEntry){
+        try{
+            bookEntry.setDate(LocalDateTime.now());
+            bookStoreRepository.save(bookEntry);
+            return new ResponseEntity<>(bookEntry,HttpStatus.CREATED);
+        }catch(Exception e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     public List<BookEntry> getAll(){
         return bookStoreRepository.findAll();
     }
 
-    public Optional<BookEntry> getById(ObjectId id){
-        return bookStoreRepository.findById(id);
+    public ResponseEntity<BookEntry> getById(ObjectId id){
+        Optional<BookEntry> book = bookStoreRepository.findById(id);
+        if(book.isPresent()){
+            return new ResponseEntity<>(book.get(),HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    public void deleteById(ObjectId id){
-        bookStoreRepository.deleteById(id);
+    public ResponseEntity<?> deleteById(ObjectId id){
+        Optional<BookEntry> book = bookStoreRepository.findById(id);
+        if(book.isPresent()){
+            bookStoreRepository.deleteById(id);
+            return new ResponseEntity<>(book.get(),HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    public BookEntry updateEntry(ObjectId id, BookEntry newEntry){
+    public ResponseEntity<BookEntry> updateEntry(ObjectId id, BookEntry newEntry){
         BookEntry old = bookStoreRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Book entry with ID " + id + " not found"));
 
         if(old != null){
@@ -42,7 +60,9 @@ public class BookStoreService {
             old.setDescription(newEntry.getDescription() != null && !newEntry.getDescription().isEmpty() ? newEntry.getDescription() : old.getDescription());
             old.setQuantity(newEntry.getQuantity() != null && newEntry.getQuantity() > 0 ? newEntry.getQuantity() : old.getQuantity());
             old.setCategoryId(newEntry.getCategoryId() != null ? newEntry.getCategoryId() : old.getCategoryId());
+            bookStoreRepository.save(old);
+            return new ResponseEntity<>(old, HttpStatus.OK);
         }
-        return bookStoreRepository.save(old);
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
